refactor(providers): clean root implementation leftovers
