+++ self/subcortex/providers/src/__tests__/chat-completions-provider.test.ts
@@ -1,8 +1,8 @@
 import { describe, it, expect, vi, beforeEach, afterEach } from 'vitest';
 import type { ProviderId } from '@nous/shared';
 import { NousError, ValidationError } from '@nous/shared';
+import { ChatCompletionsProvider as RootChatCompletionsProvider } from '../index.js';
 import { ChatCompletionsProvider } from '../protocols/openai-api/provider.js';
-import { ChatCompletionsProvider as ShimChatCompletionsProvider } from '../chat-completions-provider.js';
 
 const MOCK_CONFIG = {
   id: '00000000-0000-0000-0000-000000000001' as ProviderId,
@@ -29,8 +29,8 @@ describe('ChatCompletionsProvider', () => {
     expect(provider.getConfig()).toEqual(MOCK_CONFIG);
   });
 
-  it('keeps the root chat-completions provider path as a compatibility shim', () => {
-    expect(ShimChatCompletionsProvider).toBe(ChatCompletionsProvider);
+  it('keeps the package-root chat-completions provider export canonical', () => {
+    expect(RootChatCompletionsProvider).toBe(ChatCompletionsProvider);
   });
 
   it('constructor throws when no API key', () => {
