--- self/subcortex/providers/src/chat-completions-provider.ts
@@ -1,4 +1,0 @@
-export {
-  CHAT_COMPLETIONS_PROVIDER_DEFINITION,
-  ChatCompletionsProvider,
-} from './protocols/openai-api/provider.js';
