+++ self/subcortex/providers/src/index.ts
@@ -1,16 +1,16 @@
 /**
  * @nous/subcortex-providers — Model provider adapters for Nous-OSS.
  */
-export { AnthropicProvider } from './anthropic-provider.js';
+export { AnthropicProvider } from './providers/anthropic/implementation.js';
 export * from './adapter-registry.js';
 export * from './provider-adapters.js';
 export * from './provider-definitions.js';
 export * from './provider-factories.js';
 export * from './shared/index.js';
-export { OllamaProvider } from './ollama-provider.js';
-export { ChatCompletionsProvider } from './chat-completions-provider.js';
-export { ProviderRegistry } from './provider-registry.js';
-export type { ProviderRegistryOptions } from './provider-registry.js';
+export { OllamaProvider } from './providers/ollama/implementation.js';
+export { ChatCompletionsProvider } from './protocols/openai-api/provider.js';
+export { ProviderRegistry } from './registry/provider-registry.js';
+export type { ProviderRegistryOptions } from './registry/provider-registry.js';
 export {
   InferenceLane,
   InferenceLaneRegistry,
@@ -29,5 +29,5 @@ export type {
   ProviderBreakdownEntry,
   WindowSummary,
 } from '@nous/subcortex-inference-runtime';
-export { TextModelInputSchema } from './schemas.js';
-export type { TextModelInput } from './schemas.js';
+export { TextModelInputSchema } from './schemas/text-model-input.js';
+export type { TextModelInput } from './schemas/text-model-input.js';
