+++ self/subcortex/providers/src/providers/anthropic/definition.ts
@@ -1,3 +1,3 @@
 export {
   ANTHROPIC_PROVIDER_DEFINITION as providerDefinition,
-} from '../../anthropic-provider.js';
+} from './implementation.js';
