+++ self/subcortex/providers/src/providers/anthropic/index.ts
@@ -1,4 +1,4 @@
 export { providerAdapter } from './adapter.js';
 export { providerDefinition } from './definition.js';
 export { providerFactory } from './provider.js';
-export { AnthropicProvider } from '../../anthropic-provider.js';
+export { AnthropicProvider } from './implementation.js';
