+++ self/subcortex/providers/src/providers/ollama/definition.ts
@@ -1,3 +1,3 @@
 export {
   OLLAMA_PROVIDER_DEFINITION as providerDefinition,
-} from '../../ollama-provider.js';
+} from './implementation.js';
