+++ self/subcortex/providers/src/schemas/text-model-input.ts
